Create separate SelectDirectoryFieldEditor class.
