+++ core/ui/ui-plugins/org.csstudio.perspectives/src/org/csstudio/perspectives/PerspectivesPreferencePage.java
@@ -7,15 +7,9 @@
  ******************************************************************************/
 package org.csstudio.perspectives;
 
-import java.io.File;
-import java.io.IOException;
-
 import org.eclipse.core.runtime.preferences.InstanceScope;
 import org.eclipse.jface.preference.FieldEditorPreferencePage;
-import org.eclipse.jface.preference.StringButtonFieldEditor;
-import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.DirectoryDialog;
 import org.eclipse.ui.IWorkbench;
 import org.eclipse.ui.IWorkbenchPreferencePage;
 import org.eclipse.ui.preferences.ScopedPreferenceStore;
@@ -32,7 +26,7 @@ public class PerspectivesPreferencePage extends FieldEditorPreferencePage
 
     private IFileUtils fileUtils = new FileUtils();
 
-    private StringButtonFieldEditor perspectivesDirEditor;
+    private SelectDirectoryFieldEditor perspectivesDirEditor;
     private ScopedPreferenceStore store;
 
     public PerspectivesPreferencePage() {
@@ -44,34 +38,8 @@ public class PerspectivesPreferencePage extends FieldEditorPreferencePage
     @Override
     protected void createFieldEditors() {
         final Composite parent = getFieldEditorParent();
-        perspectivesDirEditor = new StringButtonFieldEditor(PERSPECTIVE_LOAD_DIRECTORY, Messages.PerspectivesPreferencePage_fieldText, parent) {
-            private String lastPath = store.getString(PERSPECTIVE_LOAD_DIRECTORY);
-
-            @Override
-            protected String changePressed() {
-                DirectoryDialog dialog = new DirectoryDialog(getShell(), SWT.SHEET);
-                if (lastPath != null) {
-                    try {
-                        if (new File(lastPath).exists()) {
-                            File lastDir = new File(lastPath);
-                            dialog.setFilterPath(lastDir.getCanonicalPath());
-                        }
-                    } catch (IOException e) {
-                        dialog.setFilterPath(lastPath);
-                    }
-                }
-                String dir = dialog.open();
-                if (dir != null) {
-                    String dirUri = fileUtils.stringPathToUriFileString(dir);
-                    dirUri = dirUri.trim();
-                    if (dirUri.length() == 0) {
-                        return null;
-                    }
-                    lastPath = dirUri;
-                }
-                return dir;
-            }
-        };
+        String lastPath = store.getString(PERSPECTIVE_LOAD_DIRECTORY);
+        perspectivesDirEditor = new SelectDirectoryFieldEditor(PERSPECTIVE_LOAD_DIRECTORY, Messages.PerspectivesPreferencePage_fieldText, parent, lastPath, fileUtils);
         perspectivesDirEditor.getTextControl(parent).setToolTipText(Messages.PerspectivesPreferencePage_fieldTooltip);
         addField(perspectivesDirEditor);
     }
